blog/multi-vector-embedding-knowledge-base/requirements.txt
aws-cdk-lib>=2.100.0
constructs>=10.0.0
cdklabs.generative-ai-cdk-constructs>=0.1.0
boto3>=1.34.0
psycopg2-binary>=2.9.0